refactor: Extract CertificationsIcons comp from FileName
no code change
