+++ src/modules/filelist/cells/FileName.jsx
@@ -1,13 +1,9 @@
 import cx from 'classnames'
-import get from 'lodash/get'
-import React, { useCallback } from 'react'
+import React from 'react'
 import { Link } from 'react-router-dom'
 
-import { useClient } from 'cozy-client'
 import { isDirectory } from 'cozy-client/dist/models/file'
-import AppIcon from 'cozy-ui/transpiled/react/AppIcon'
 import Icon from 'cozy-ui/transpiled/react/Icon'
-import CarbonCopyIcon from 'cozy-ui/transpiled/react/Icons/CarbonCopy'
 import ShareIcon from 'cozy-ui/transpiled/react/Icons/Share'
 import MidEllipsis from 'cozy-ui/transpiled/react/MidEllipsis'
 import { TableCell } from 'cozy-ui/transpiled/react/deprecated/Table'
@@ -16,66 +12,12 @@ import { useI18n } from 'cozy-ui/transpiled/react/providers/I18n'
 import styles from '@/styles/filelist.styl'
 
 import RenameInput from '@/modules/drive/RenameInput'
+import CertificationsIcons from '@/modules/filelist/cells/CertificationsIcons'
 import {
   getFileNameAndExtension,
   makeParentFolderPath
 } from '@/modules/filelist/helpers'
 import { isSharedDriveFolder } from '@/modules/shareddrives/helpers'
-
-export const CertificationsIcons = ({ attributes }) => {
-  const isCarbonCopy = get(attributes, 'metadata.carbonCopy')
-  const isElectronicSafe = get(attributes, 'metadata.electronicSafe')
-  const slug = get(attributes, 'cozyMetadata.uploadedBy.slug')
-  const client = useClient()
-
-  // TODO To be removed when UI's AppIcon use getIconURL from Cozy-Client
-  // instead of its own see https://github.com/cozy/cozy-ui/issues/1723
-  const fetchIcon = useCallback(() => {
-    return client.getStackClient().getIconURL({
-      type: 'konnector',
-      slug,
-      priority: 'registry'
-    })
-  }, [client, slug])
-
-  return (
-    <div className={styles['fil-file-certifications']}>
-      {(isCarbonCopy || isElectronicSafe) && (
-        <span className={styles['fil-file-certifications--separator']}>
-          {' - '}
-        </span>
-      )}
-      {isCarbonCopy &&
-        (isElectronicSafe ? (
-          <span data-testid="certificationsIcons-carbonCopyIcon">
-            <Icon
-              icon={CarbonCopyIcon}
-              className={`u-mr-half ${styles['fil-file-certifications--icon']}`}
-            />
-          </span>
-        ) : (
-          <span data-testid="certificationsIcons-carbonCopyAppIcon">
-            <AppIcon
-              app={slug}
-              className={styles['fil-file-certifications--icon']}
-              fetchIcon={fetchIcon}
-              type="konnector"
-            />
-          </span>
-        ))}
-      {isElectronicSafe && (
-        <span data-testid="certificationsIcons-electronicSafeAppIcon">
-          <AppIcon
-            app={slug}
-            className={styles['fil-file-certifications--icon']}
-            fetchIcon={fetchIcon}
-            type="konnector"
-          />
-        </span>
-      )}
-    </div>
-  )
-}
 
 const FileName = ({
   attributes,
